requirements.txt
Django
django-widget-tweaks
djangorestframework == 3.12.2
Pillow == 8.0.1
django-filter == 2.4.0
drf-yasg